requirements.txt
streamlit
streamlit_javascript
rembg
requests
pillow
